+++ tests/end_to_end/test_question_service.py
@@ -8,7 +8,7 @@ async def test_all_crud_question_no_errors(question_service):
     """ Test a standard CRUD operation flow with no errors from the user end """
 
     data = {"source": "other", "link": "", "difficulty": "easy", "status": "completed", "title": "All CRUD end to end",
-            "prompt": "no user errors", "test_cases": [], "notes": [], "hints": [], "tags": []}
+            "prompt": "no user errors", "notes": [], "hints": [], "tags": []}
     q = await question_service.create_question(QuestionCreate(**data))
     id = q.id
 
@@ -40,7 +40,6 @@ async def test_deleting_all_questions(question_service):
 
     # Insert back to database for other tests
     for q in deleted:
-        data = {"source": q.source, "link": q.link, "difficulty": q.difficulty, "status": q.status, "title": q.title,
-                "prompt": q.prompt, "test_cases": q.test_cases, "notes": q.notes, "hints": q.hints,
-                "tags": q.tags}
+        data = {"source": q.source, "link": q.link, "difficulty": q.difficulty, "status": q.status,
+                "title": q.title, "prompt": q.prompt, "notes": q.notes, "hints": q.hints, "tags": q.tags}
         await question_service.create_question(QuestionCreate(**data), q.id)